Fix foreign key data type mismatch for loans table migration
- Update `database/migrations/2024_01_02_000000_create_pinjaman_table.php` to use `unsignedBigInteger` for `anggota_id`, matching the `bigIncrements` (BigInt) primary key of the `anggota` table.
- This resolves the `General error: 1215 Cannot add foreign key constraint` during migration.

+++ database/migrations/2024_01_02_000000_create_pinjaman_table.php
@@ -16,7 +16,7 @@ class CreatePinjamanTable extends Migration
         Schema::create('pinjaman', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('kode_pinjaman')->unique();
-            $table->unsignedInteger('anggota_id');
+            $table->unsignedBigInteger('anggota_id');
             $table->string('jenis_pinjaman'); // produktif, konsumtif
             $table->decimal('jumlah_pinjaman', 15, 2);
             $table->integer('tenor'); // in months
